Move token conversion into it's own module
